halfway through form generation rework
committing for the night. lots of files here that will be deleted once done.
current state: generates general and training. need to re-connect the interaction but should be much clearer than it used to be. four other tabs need to be made - need to develop a design for (h)param handling first.